requirements.txt
onnxruntime==1.20.1
onnxruntime-gpu==1.18.0
insightface==0.7.3
boto3==1.34.34
python-dotenv==1.0.1
requests==2.31.0
psycopg2-binary==2.9.9
google-generativeai==0.3.2
Pillow==10.2.0